All the EventTasks work now!

+++ app/src/main/java/com/github/funnygopher/parti/invitation/AttendEventTask.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 public class AttendEventTask extends AsyncTask<Void, Void, String> {
 
     private static final String ADDRESS = "http://pumpuptheparti.netne.net/api/attend_event.php";
-    private int eventId;
+    private Long eventId;
     private OnResponseListener responseListener = new OnResponseListener() {
         @Override
         public void onResponse(String response) {
@@ -18,7 +18,7 @@ public class AttendEventTask extends AsyncTask<Void, Void, String> {
         }
     };
 
-    public AttendEventTask(int eventId) {
+    public AttendEventTask(Long eventId) {
         this.eventId = eventId;
     }
 
@@ -26,7 +26,7 @@ public class AttendEventTask extends AsyncTask<Void, Void, String> {
     protected String doInBackground(Void... params) {
         try {
             HttpRequest httpRequest = new HttpRequest(HttpRequest.POST, ADDRESS);
-            httpRequest.withString("id=" + Integer.toString(eventId));
+            httpRequest.withString("id=" + Long.toString(eventId));
             String response = httpRequest.send();
             Log.i("AttendEvent#Response", response);
             return response;
